feat: add carrier and genome crates, migrate VFS to genome
- Added new workspace crates `carrier` (spinlock utilities) and `genome` (VFS implementation)
- Bumped `spin` dependency from 0.9.8 to 0.10.0 across the workspace
- Refactored `nozzle` to use `carrier` for synchronization primitives
- Extracted VFS module from `nozzle` into `genome`, integrated into kernel and nozzle
- Removed inline documentation from `nozzle` VFS wrappers in favor of canonical source
