+++ carrier/src/lib.rs
@@ -1,0 +1,6 @@
+#![no_std]
+extern crate alloc;
+
+pub mod terminal;
+pub mod pipeline;
+pub mod exec;
