+++ carrier/src/pipeline.rs
@@ -1,25 +1,14 @@
-//! Simple command-line parser for Nozzle
-//!
-//! Splits input into command name and arguments.
-//! Supports pipe (`|`) chaining and basic quoting.
-
 use alloc::fmt;
 use alloc::string::String;
 use alloc::string::ToString;
 use alloc::vec::Vec;
 
-/// A pipeline of commands connected by `|`.
-///
-/// Example: `ls | grep foo | wc -l` produces three [`ParsedCommand`]s.
 #[derive(Debug, Clone)]
 pub struct Pipeline {
     pub commands: Vec<ParsedCommand>,
 }
 
 impl Pipeline {
-    /// Parse a line that may contain pipes.
-    ///
-    /// The input should already be trimmed of leading/trailing whitespace.
     pub fn parse(line: &str) -> Self {
         let commands: Vec<ParsedCommand> = line
             .split('|')
@@ -29,7 +18,6 @@ impl Pipeline {
         Self { commands }
     }
 
-    /// `true` when the pipeline has exactly one command (no pipe).
     pub fn is_simple(&self) -> bool {
         self.commands.len() <= 1
     }
@@ -47,19 +35,13 @@ impl fmt::Display for Pipeline {
     }
 }
 
-/// A parsed command
 #[derive(Debug, Clone)]
 pub struct ParsedCommand {
-    /// The command name (first token)
     pub name: String,
-    /// Arguments (remaining tokens)
     pub args: Vec<String>,
 }
 
 impl ParsedCommand {
-    /// Parse a command line string.
-    ///
-    /// The input should already be trimmed of leading/trailing whitespace.
     pub fn parse(line: &str) -> Self {
         let mut parts = line.split_whitespace();
         let name = parts.next().unwrap_or("").to_string();
@@ -67,7 +49,6 @@ impl ParsedCommand {
         Self { name, args }
     }
 
-    /// Get all tokens as `&[&str]` (args[0] is the command name).
     pub fn args_slice(&self) -> Vec<&str> {
         let mut result = Vec::with_capacity(self.args.len() + 1);
         result.push(&*self.name);
