+++ genome/src/lib.rs
@@ -1,0 +1,5 @@
+#![no_std]
+extern crate alloc;
+
+pub mod vfs;
+pub mod fs;
