+++ nozzle/src/lib.rs
@@ -1,46 +1,26 @@
-//! Nozzle — interactive shell runtime for Fullerene OS
-//!
-//! Nozzle is a no_std interactive shell runtime that provides:
-//!
-//! - A [`Terminal`] trait for abstract I/O
-//! - A [`LineEditor`] with history, cursor movement, and editing
-//! - A [`Command`] trait for extensible built-in commands
-//! - Built-in commands: `clear`, `echo`, `exit`, `uname`
-//! - `help` is handled by the dispatch function directly
-//! - A main loop that ties everything together
-//!
-//! Nozzle knows nothing about framebuffers, graphics, or the kernel
-//! — it only needs a text-based I/O stream.
-
 #![no_std]
 extern crate alloc;
 
 pub mod builtins;
-pub mod exec;
 pub mod fs_hooks;
 pub mod line_editor;
-pub mod parser;
 pub mod prompt;
 pub mod selection;
 pub mod sys_hooks;
-pub mod terminal;
 pub mod terminal_buffer;
 
 use alloc::string::String;
 
-pub use exec::{Command, CommandContext, NamedCommand};
-pub use line_editor::LineEditor;
-pub use parser::ParsedCommand;
-pub use prompt::Prompt;
-pub use terminal::Terminal;
+// Re-export carrier types so existing consumers still work
+pub use carrier::exec::{Command, CommandContext, NamedCommand};
+pub use carrier::pipeline::ParsedCommand;
+pub use carrier::terminal::Terminal;
 
-/// Default shell prompt
+pub use line_editor::LineEditor;
+pub use prompt::Prompt;
+
 pub const DEFAULT_PROMPT: &str = "nozzle> ";
 
-/// The Nozzle shell runtime.
-///
-/// Wires together a [`Terminal`], [`LineEditor`], and command list
-/// into an interactive read-eval-print loop.
 pub struct Shell<'a> {
     terminal: &'a mut dyn Terminal,
     commands: &'a [&'a dyn Command],
@@ -50,7 +30,6 @@ pub struct Shell<'a> {
 }
 
 impl<'a> Shell<'a> {
-    /// Create a new shell instance.
     pub fn new(terminal: &'a mut dyn Terminal, commands: &'a [&'a dyn Command]) -> Self {
         Self {
             terminal,
@@ -61,26 +40,19 @@ impl<'a> Shell<'a> {
         }
     }
 
-    /// Set the prompt string.
     pub fn set_prompt(&mut self, text: impl Into<String>) {
         self.prompt.set_text(text);
     }
 
-    /// Run the main shell loop.
-    ///
-    /// This function returns when the `exit` command is executed.
     pub fn run(&mut self) {
         self.show_welcome();
 
         loop {
-            // Show prompt
             self.terminal.write_str(self.prompt.as_str());
 
-            // Read a line
             let line = match self.editor.read_line(&mut *self.terminal) {
                 Some(l) => l,
                 None => {
-                    // Ctrl+C or Ctrl+D on empty line
                     continue;
                 }
             };
@@ -90,8 +62,7 @@ impl<'a> Shell<'a> {
                 continue;
             }
 
-            // Dispatch and execute
-            let should_continue = exec::dispatch(self.commands, &mut *self.terminal, trimmed);
+            let should_continue = carrier::exec::dispatch(self.commands, &mut *self.terminal, trimmed);
             if !should_continue {
                 break;
             }
@@ -111,12 +82,9 @@ impl<'a> Shell<'a> {
     }
 }
 
-/// Create the default command list (excluding `help`, which is handled by `dispatch`).
-///
-/// Returns a `&'static` slice suitable for passing to [`Shell::new`].
 pub fn default_commands() -> &'static [&'static dyn Command] {
     use crate::builtins;
-    define_commands!(
+    carrier::define_commands!(
         ("clear", "Clear the screen", builtins::cmd_clear),
         ("echo", "Print text", builtins::cmd_echo),
         ("exit", "Exit the shell", builtins::cmd_exit),
@@ -134,39 +102,19 @@ pub fn default_commands() -> &'static [&'static dyn Command] {
         ("shutdown", "Shutdown the system", builtins::cmd_shutdown),
         ("calc", "Simple arithmetic calculator", builtins::cmd_calc),
         ("run", "Launch an external application", builtins::cmd_run),
-        (
-            "taskmon",
-            "Detailed task/process monitor",
-            builtins::cmd_taskmon
-        ),
-        (
-            "devices",
-            "List registered hardware devices",
-            builtins::cmd_devices
-        ),
+        ("taskmon", "Detailed task/process monitor", builtins::cmd_taskmon),
+        ("devices", "List registered hardware devices", builtins::cmd_devices),
         ("theme", "Show or change desktop theme", builtins::cmd_theme),
-        (
-            "wallpaper",
-            "Show or change desktop wallpaper",
-            builtins::cmd_wallpaper
-        ),
+        ("wallpaper", "Show or change desktop wallpaper", builtins::cmd_wallpaper),
         ("pci", "List PCI devices", builtins::cmd_pci),
-        (
-            "badapple",
-            "Play Bad Apple!! animation",
-            builtins::cmd_badapple
-        ),
+        ("badapple", "Play Bad Apple!! animation", builtins::cmd_badapple),
         ("cd", "Change working directory", builtins::cmd_cd),
         ("tree", "Display directory tree", builtins::cmd_tree),
         ("find", "Search for files", builtins::cmd_find),
         ("cp", "Copy a file", builtins::cmd_cp),
         ("mv", "Move a file", builtins::cmd_mv),
         ("write", "Write content to a file", builtins::cmd_write),
-        (
-            "app",
-            "Package manager (install/remove/list)",
-            builtins::cmd_app
-        ),
+        ("app", "Package manager (install/remove/list)", builtins::cmd_app),
         ("rm", "Remove files or directories", builtins::cmd_rm),
         ("mkdir", "Create directories", builtins::cmd_mkdir),
         ("touch", "Create empty files", builtins::cmd_touch),
@@ -179,20 +127,12 @@ pub fn default_commands() -> &'static [&'static dyn Command] {
         ("grep", "Search for a pattern", builtins::cmd_grep),
         ("sort", "Sort lines of text", builtins::cmd_sort),
         ("wc", "Count lines, words, and bytes", builtins::cmd_wc),
-        (
-            "hello_linux",
-            "Launch the built-in Linux test binary",
-            builtins::cmd_hello_linux
-        ),
-        (
-            "linux_run",
-            "Launch a Linux ELF binary from the filesystem",
-            builtins::cmd_linux_run
-        ),
-        (
-            "run_busybox",
-            "Launch BusyBox shell from the filesystem",
-            builtins::cmd_run_busybox
-        ),
+        ("hello_linux", "Launch the built-in Linux test binary", builtins::cmd_hello_linux),
+        ("linux_run", "Launch a Linux ELF binary from the filesystem", builtins::cmd_linux_run),
+        ("run_busybox", "Launch BusyBox shell from the filesystem", builtins::cmd_run_busybox),
     )
+}
+
+pub fn get_completions(prefix: &str) -> alloc::vec::Vec<alloc::string::String> {
+    carrier::exec::get_completions_for(prefix, default_commands())
 }
