--- nozzle/src/terminal.rs
@@ -1,54 +1,0 @@
-//! Terminal abstraction for Nozzle shell
-//!
-//! This trait decouples the shell from any specific I/O backend.
-//! The kernel (or any other environment) provides a concrete implementation.
-
-/// Abstract terminal I/O interface
-pub trait Terminal {
-    /// Write a string to the terminal
-    fn write_str(&mut self, s: &str);
-
-    /// Read a single byte from input, blocking until available.
-    /// Returns `None` on end-of-input.
-    fn read_byte(&mut self) -> Option<u8>;
-
-    /// Check if input data is available without blocking
-    fn input_available(&self) -> bool {
-        false
-    }
-
-    /// Feed a string as stdin (e.g. from a pipe).
-    ///
-    /// The next call to [`read_byte`] should consume from this buffer
-    /// before polling the real input source.
-    fn set_stdin(&mut self, _data: alloc::string::String) {}
-
-    /// Collect stdout accumulated since the last call.
-    ///
-    /// Used by the pipe dispatcher to capture a command's output.
-    /// Returns `None` if no output was produced.
-    fn take_stdout(&mut self) -> Option<alloc::string::String> {
-        None
-    }
-
-    /// Take stdin data previously set by [`set_stdin`] for consumption
-    /// by a command (e.g. `grep`, `sort`, `wc` when receiving pipe input).
-    ///
-    /// Returns `Some(String)` if stdin data is available, `None` otherwise.
-    /// Consuming the data clears the internal buffer.
-    fn take_stdin(&mut self) -> Option<alloc::string::String> {
-        None
-    }
-
-    /// Arm the pipe stdout buffer to capture output for the current stage.
-    ///
-    /// Called before each command in a pipeline executes to ensure
-    /// that `write_str` output is captured into the pipe buffer.
-    fn arm_pipe_stdout(&mut self) {}
-
-    /// Clear any residual pipe stdin data after a stage completes.
-    ///
-    /// Called after each command in a pipeline to ensure stale input
-    /// doesn't leak into the next stage.
-    fn clear_pipe_stdin(&mut self) {}
-}
